refactor: 상태 패턴 개선 및 Stay 판정 로직 분리
  - State 인터페이스에서 hand() 제거, score()/isBust()/cards() 추가
  - profit 타입 double → BigDecimal 변경
  - Stay를 Finished에서 분리, judge(dealerScore)로 Win/Lose/Draw 전환
  - Win/Lose/Draw 상태 추가
  - Hit.draw()에서 불필요한 List 래핑 제거

+++ src/main/java/domain/state/Running.java
@@ -2,8 +2,10 @@ package domain.state;
 
 import domain.Hand;
 
-public abstract class Running extends Started{
-   protected Running(Hand hand) {
+import java.math.BigDecimal;
+
+public abstract class Running extends Started {
+    protected Running(Hand hand) {
         super(hand);
     }
 
@@ -13,7 +15,7 @@ public abstract class Running extends Started{
     }
 
     @Override
-    public double profit(double betAmount) {
+    public BigDecimal profit(BigDecimal betAmount) {
         throw new IllegalStateException("아직 진행중입니다.");
     }
 }
